# server/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"tasks": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"tasks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"token": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"status"
				],
				"type": "object",
				"description": "Auth JSON response for Controllers"
			},
			"Auth": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Task": {
				"properties": {
					"description": {
						"type": "string"
					},
					"completed": {
						"type": "boolean"
					}
				},
				"required": [
					"description",
					"completed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserResponse": {
				"properties": {
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"status"
				],
				"type": "object",
				"description": "User JSON response for Controllers"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "task-project-backend",
		"version": "1.0.0",
		"description": "Backend for Task Project",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Enrique"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/auth/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "User registered succeful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"404": {
						"description": "The email is already in use"
					}
				},
				"description": "Endpoint to Register new user",
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Data of user",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User",
								"description": "Data of user"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Promise of AuthResponse with a JWT",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"description": "Endpoint to log in user",
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "User credentials",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Auth",
								"description": "User credentials"
							}
						}
					}
				}
			}
		},
		"/api/v1/me": {
			"get": {
				"operationId": "UserData",
				"responses": {
					"200": {
						"description": "Returns Promise of User data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"description": "Endpoint to get User data after log in",
				"tags": [
					"UserController"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of User by Query param /me?id=ID",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/me/tasks": {
			"post": {
				"operationId": "SaveTasks",
				"responses": {
					"200": {
						"description": "Promise of tasks saved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"description": "Endpoint to save the user's tasks",
				"tags": [
					"UserController"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of User by Query param /me/tasks?id=ID",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "User tasks",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/Task"
								},
								"type": "array",
								"description": "User tasks"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}